Split spectroscopy out of plans.py
