+++ src/test_rig_bluesky/plans.py
@@ -6,26 +6,18 @@ from typing import Any
 import bluesky.plan_stubs as bps
 import bluesky.preprocessors as bpp
 from bluesky.plans import count
-from bluesky.protocols import Movable
+from bluesky.protocols import Movable, Readable
 from bluesky.utils import MsgGenerator
 from dodal.common import inject
-from dodal.devices.motors import XYZStage
 from dodal.plan_stubs.data_session import attach_data_session_metadata_decorator
-from dodal.plans import spec_scan
 from ophyd_async.core import (
     DetectorTrigger,
     Device,
     Settings,
     SettingsProvider,
+    StandardDetector,
     TriggerInfo,
     YamlSettingsProvider,
-)
-from ophyd_async.epics.adaravis import AravisDetector
-from ophyd_async.epics.adcore import (
-    NDAttributeDataType,
-    NDAttributeParam,
-    NDROIStatNIO,
-    setup_ndattributes,
 )
 from ophyd_async.epics.pmac import PmacIO, PmacScanInfo, PmacTrajectoryFlyableLogic
 from ophyd_async.fastcs.panda import (
@@ -41,16 +33,15 @@ from ophyd_async.plan_stubs import (
     retrieve_settings,
     store_settings,
 )
-from scanspec.specs import Fly, Line, Spec
+from pydantic import TypeAdapter
+from scanspec.specs import Fly, Spec
 
 LOGGER = logging.getLogger(__name__)
 
-# imaging_detector = inject("imaging_detector")
-spectroscopy_detector = inject("spectroscopy_detector")
-sample_stage = inject("sample_stage")
+# Shared hardware only. Experiment-specific devices are injected in the module
+# for that experiment, so that this one can be imported by both.
 pmac = inject("pmac")
 pandabrick = inject("pandabrick")
-# pmac_trigger_logic = inject("pmac_trigger_logic")
 
 
 def save_settings(
@@ -104,204 +95,79 @@ def _settings_provider() -> SettingsProvider:
     return YamlSettingsProvider(this_directory)
 
 
+def serialize_spec(spec: Spec[Any]) -> Any:
+    """Render a Spec as something JSON-serializable, for scan metadata."""
+    # NOTE: replace with spec.serialize() when this merges:
+    # https://github.com/bluesky/scanspec/pull/208
+    return TypeAdapter(Spec[Any]).dump_python(spec, mode="json", fallback=repr)
+
+
 @attach_data_session_metadata_decorator()
-def snapshot(
-    imaging_detector: AravisDetector,
-    spectroscopy_detector: AravisDetector = spectroscopy_detector,
-    sample_stage: XYZStage = sample_stage,
-) -> MsgGenerator[None]:
-    """Capture a snapshot of the current state of the beamline."""
-    yield from count([imaging_detector, spectroscopy_detector, sample_stage])
+def snapshot(devices: list[Readable]) -> MsgGenerator[None]:
+    """Capture a snapshot of the current state of the given devices.
 
-
-def spectroscopy(
-    spectroscopy_detector: AravisDetector = spectroscopy_detector,
-    sample_stage: XYZStage = sample_stage,
-    pmac: PmacIO = pmac,
-    pandabrick: HDFPanda = pandabrick,
-    num_points: int = 25,
-    spec: Spec[Movable] | None = None,
-    exposure_time: float = 0.1,
-    fly: bool = True,
-    metadata: dict[str, Any] | None = None,
-) -> MsgGenerator[None]:
-    """Do a spectroscopy scan."""
-    yield from load_settings(
-        device=spectroscopy_detector,
-        design_name="spectroscopy_detector_baseline",
-        whitelist_pvs=[
-            "hdf-nd_array_port",
-            "hdf-enable_callbacks",
-            "driver-acquire",
-            "driver-trigger_mode",
-            "driver-trigger_source",
-            "roistat-channels-array_counter",
-            "roistat-channels-1-min_x",
-            "roistat-channels-1-min_y",
-            "roistat-channels-1-name_",
-            "roistat-channels-1-size_x",
-            "roistat-channels-1-size_y",
-            "roistat-channels-1-use",
-            "roistat-channels-2-min_x",
-            "roistat-channels-2-min_y",
-            "roistat-channels-2-name_",
-            "roistat-channels-2-size_x",
-            "roistat-channels-2-size_y",
-            "roistat-channels-2-use",
-            "roistat-channels-3-min_x",
-            "roistat-channels-3-min_y",
-            "roistat-channels-3-name_",
-            "roistat-channels-3-size_x",
-            "roistat-channels-3-size_y",
-            "roistat-channels-3-use",
-            "roistat-nd_array_port",
-            "roistat-enable_callbacks",
-        ],
-    )
-
-    # We call mv instead of prepare because prepare cannot technically be used
-    # outside of a run.
-    # See: https://github.com/DiamondLightSource/blueapi/issues/1211
-    #
-    # Deadtime taken from
-    # https://github.com/bluesky/ophyd-async/blob/15fa34b6ea2a28e2f27265a5564c9ee36423f1b7/src/ophyd_async/epics/adaravis/_aravis_controller.py#L11
-    yield from bps.mv(
-        *(spectroscopy_detector.driver.acquire_time, exposure_time),
-        *(spectroscopy_detector.driver.acquire_period, exposure_time + 1961e-6),
-        group="spectroscopy_detector_aquire",
-    )
-
-    params: list[NDAttributeParam] = []
-    for channel in list(spectroscopy_detector.roistat.channels.keys()):  # type: ignore
-        roistatn = spectroscopy_detector.roistat.channels[channel]  # type: ignore
-        assert isinstance(roistatn, NDROIStatNIO)
-
-        channel_name = yield from bps.rd(roistatn.name_)
-
-        params.append(
-            NDAttributeParam(
-                name=f"{channel_name}Total",
-                param="ROISTAT_TOTAL",
-                datatype=NDAttributeDataType.DOUBLE,
-                addr=channel - 1,
-                description=f"Sum of {channel_name} channel",
-            )
-        )
-
-    yield from setup_ndattributes(spectroscopy_detector.roistat, params)  # type: ignore
-
-    yield from load_settings(
-        device=sample_stage,
-        design_name="sample_stage_baseline",
-        whitelist_pvs=[
-            "x-acceleration_time",
-            "x-velocity",
-            "y-acceleration_time",
-            "y-velocity",
-        ],
-    )
-
-    # yield from load_panda_settings(
-    #     panda=pandabrick,
-    #     design_name="pandabrick_baseline",
-    #     whitelist_pvs=["incenc-__3-val_dataset", "incenc-__2-val_dataset"],
-    # )
-
-    spec = spec or Line(sample_stage.x, 0, 5, 5)  # type: ignore
-
-    # NOTE: replace with spec_serialized = spec.serialize() when this merges: https://github.com/bluesky/scanspec/pull/208
-    from typing import Any
-
-    from pydantic import TypeAdapter
-    from scanspec.specs import Spec
-
-    spec_serialzied = TypeAdapter(Spec[Any]).dump_python(
-        spec, mode="json", fallback=repr
-    )
-
-    # write json-serializable form of spec into metadata
-    if metadata is None:
-        metadata = {}
-    # spec.serialize() when https://github.com/bluesky/scanspec/pull/208 merges
-    metadata["spec"] = spec_serialzied
-
-    if fly:
-        LOGGER.info("Performing a fly scan.")
-        yield from fly_scan(
-            spec,
-            spectroscopy_detector,
-            sample_stage,
-            pmac,
-            pandabrick,
-            num_points,  # type: ignore
-            exposure_time,
-            metadata,
-        )
-    else:
-        LOGGER.info("Performing a step scan.")
-        yield from spec_scan(
-            {spectroscopy_detector, sample_stage},
-            spec,  # type: ignore
-            metadata=metadata,
-        )
+    Useful as a smoke test for new hardware: it exercises device connection and
+    the data-writing path without any triggering complexity.
+    """
+    yield from count(devices)
 
 
 def fly_scan(
     spec: Spec[Movable],
-    spectroscopy_detector: AravisDetector = spectroscopy_detector,
-    sample_stage: XYZStage = sample_stage,
+    detector: StandardDetector,
     pmac: PmacIO = pmac,
     pandabrick: HDFPanda = pandabrick,
-    num_points: int = 1_000,
     exposure_time: float = 0.1,
+    detector_deadtime: float = 2e-3 * 1.01,
+    detector_trigger: DetectorTrigger = DetectorTrigger.EXTERNAL_EDGE,
+    panda_whitelist: list[str] | None = None,
+    plan_name: str = "fly_scan",
     metadata: dict[str, Any] | None = None,
 ):
-    # yield from load_panda_settings(panda=pandabrick,design_name="pandabrick_baseline")
-
     yield from load_panda_settings(
         panda=pandabrick,
         design_name="pandabrick_baseline",
-        whitelist_pvs=["incenc-3-val_dataset", "incenc-2-val_dataset"],
+        whitelist_pvs=panda_whitelist,
     )
 
     scan_frame_duration = exposure_time
     fly_spec = Fly(scan_frame_duration @ spec)  # type: ignore
-    detector_deadtime = 2e-3 * 1.01
+
+    # Derived from the spec rather than passed in: if the two disagree the scan
+    # either truncates or hangs waiting for frames that never arrive.
+    num_points = math.prod(fly_spec.shape())
 
     table: SeqBlock = pandabrick.seq[1]  # type: ignore # noqa: SLF001
 
     pmac_scan_info = PmacScanInfo(spec=fly_spec, ramp_time=None, turnaround_time=None)  # type: ignore
     scan_spec_info = ScanSpecInfo(spec=fly_spec, deadtime=detector_deadtime)  # type: ignore
 
-    # motor_pos_out = {
-    #     sample_stage.x: PosOutScaleOffset.from_inenc(pandabrick, 2),
-    #     sample_stage.y: PosOutScaleOffset.from_inenc(pandabrick, 3)
-    # }
-
     # The trajectory and seq-table logics are bare FlyableLogic; wrap each in an
     # ephemeral StandardFlyable so the RunEngine can prepare/kickoff/complete it.
     pmac_trigger_logic = PmacTrajectoryFlyableLogic(pmac).with_device(
         name="pmac_trigger_logic"
     )
-    panda_trigger_logic = ScanSpecSeqTableFlyableLogic(
-        table,
-        # motor_pos_out
-    ).with_device(name="panda_trigger_logic")
+    panda_trigger_logic = ScanSpecSeqTableFlyableLogic(table).with_device(
+        name="panda_trigger_logic"
+    )
 
     scan_frame_livetime = scan_frame_duration - detector_deadtime
 
     # Reproduce metadata present in spec_scan
     _md = {
         "plan_args": {
-            "detectors": {det.name for det in [pandabrick, spectroscopy_detector]},
+            "detectors": {det.name for det in [pandabrick, detector]},
             "spec": repr(fly_spec),
         },
-        "plan_name": "fly_scan",
+        "plan_name": plan_name,
         "shape": fly_spec.shape(),
         **(metadata or {}),
     }
 
-    # Prepare Panda file writer trigger info
+    # Prepare Panda file writer trigger info.
+    # The PandA HDF writer takes EXTERNAL_LEVEL and the camera EXTERNAL_EDGE;
+    # the difference is deliberate. Getting these wrong causes silent
+    # frame-count mismatches.
     panda_hdf_info = TriggerInfo(
         number_of_events=num_points,
         trigger=DetectorTrigger.EXTERNAL_LEVEL,
@@ -309,10 +175,9 @@ def fly_scan(
         deadtime=detector_deadtime,
     )
 
-    # Prepare Panda file writer trigger info
     detector_info = TriggerInfo(
         number_of_events=num_points,
-        trigger=DetectorTrigger.EXTERNAL_EDGE,
+        trigger=detector_trigger,
         livetime=scan_frame_livetime,
         deadtime=detector_deadtime,
     )
@@ -320,25 +185,22 @@ def fly_scan(
     @attach_data_session_metadata_decorator()
     @bpp.run_decorator(md=_md)
     @bpp.stage_decorator(
-        [pandabrick, panda_trigger_logic, spectroscopy_detector, pmac_trigger_logic]
+        [pandabrick, panda_trigger_logic, detector, pmac_trigger_logic]
     )
     def inner_plan():
-        # create a group that is waited for
-        # Hashable prepare_group = ["pmac_trigger_logic", "trigger_logic"]
-
         # Prepare pmac with the trajectory
         yield from bps.prepare(pmac_trigger_logic, pmac_scan_info)
         # prepare sequencer table
         yield from bps.prepare(panda_trigger_logic, scan_spec_info)
         # prepare panda and hdf writer once, at start of scan
         yield from bps.prepare(pandabrick, panda_hdf_info)
-        # prepare spectroscopy_detector and info
+        # prepare detector and info
         # waiting for this last prepare means all prepare functions will be complete
-        yield from bps.prepare(spectroscopy_detector, detector_info, wait=True)
+        yield from bps.prepare(detector, detector_info, wait=True)
 
         # Need to run this after detectors are prepared
         yield from bps.declare_stream(
-            pandabrick, spectroscopy_detector, name="primary", collect=True
+            pandabrick, detector, name="primary", collect=True
         )
 
         # Start the detectors and hdf writers acquiring.
@@ -347,7 +209,7 @@ def fly_scan(
         # its own with wait=true.
         yield from bps.kickoff(pandabrick)
         yield from bps.kickoff(panda_trigger_logic)
-        yield from bps.kickoff(spectroscopy_detector, wait=True)
+        yield from bps.kickoff(detector, wait=True)
 
         # Start the trajectory.
         yield from bps.kickoff(pmac_trigger_logic, wait=True)
@@ -358,75 +220,14 @@ def fly_scan(
                 pmac_trigger_logic,
                 panda_trigger_logic,
                 pandabrick,
-                spectroscopy_detector,
+                detector,
             ),
             dets=(
                 pandabrick,
-                spectroscopy_detector,
+                detector,
             ),
             stream_name="primary",
             flush_period=0.5,
         )
 
     yield from inner_plan()
-
-
-def demo_spectroscopy(
-    spectroscopy_detector: AravisDetector = spectroscopy_detector,
-    sample_stage: XYZStage = sample_stage,
-    pmac: PmacIO = pmac,
-    pandabrick: HDFPanda = pandabrick,
-    total_number_of_scan_points: int = 25,
-    grid_size: float = 5.0,
-    grid_origin_x: float = 0.0,
-    grid_origin_y: float = 0.0,
-    exposure_time: float = 0.1,
-    metadata: dict[str, Any] | None = None,
-) -> MsgGenerator[None]:
-    """Spectroscopy plan intended for use in Visr demonstrations to visitors.
-    The time taken to scan is approximately linear in total_numbers_of_grid_points.
-    All other parameters can be left at their defaults.
-    """
-    xsteps = ysteps = int(round(math.sqrt(max(total_number_of_scan_points, 1))))
-    xmin = grid_origin_x
-    xmax = grid_origin_x + grid_size
-    ymin = grid_origin_y
-    ymax = grid_origin_y + grid_size
-
-    velomax = grid_size / (xsteps * exposure_time)
-    if velomax <= 10 * 0.98:
-        fly = True
-        LOGGER.info("test-rig-bluesky")
-        LOGGER.info(
-            f"Estimated velocity is {velomax:.2f} mm/sec, performing a fly scan."
-        )
-    else:
-        fly = False
-        LOGGER.info(
-            f"Estimated velocity is {velomax:.2f} mm/sec, performing a step scan."
-        )
-
-    # Move to the start point
-    # pmac should handle premove
-    yield from bps.mv(
-        *(sample_stage.x, xmin), *(sample_stage.y, ymin), group="initial_move"
-    )
-
-    grid = Line(sample_stage.y, ymin, ymax, ysteps) * ~Line(  # type: ignore
-        sample_stage.x,  # type: ignore
-        xmin,
-        xmax,
-        xsteps,
-    )
-    # fly = False
-    yield from spectroscopy(
-        spectroscopy_detector=spectroscopy_detector,
-        sample_stage=sample_stage,
-        pmac=pmac,
-        pandabrick=pandabrick,
-        num_points=total_number_of_scan_points,
-        spec=grid,
-        exposure_time=exposure_time,
-        fly=fly,
-        metadata=metadata,
-    )
